moved ShaderPreprocessor into ci::gl namespace
